requirements.txt
streamlit==1.26.1
opencv-python-headless
numpy
pillow
streamlit-drawable-canvas
